Conflicts:
	src/include/planner/plan_util.h

+++ test/type/timestamp_value_test.cpp
@@ -97,7 +97,7 @@ TEST_F(TimestampValueTests, ComparisonTest) {
         }  // SWITCH
         LOG_TRACE("%s %s %s => %d | %d\n", val0.ToString().c_str(),
                   ExpressionTypeToString(etype).c_str(),
-                  val1.ToString().c_str(), expected, result);
+                  val1.ToString().c_str(), expected, static_cast<int>(result));
 
         if (expected_null) {
           EXPECT_EQ(expected_null, result == type::CmpBool::NULL_);
